# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing passengers and taxists in a ride-sharing application.",
        "title": "Ride-Sharing Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.55.42:8080",
    "basePath": "/",
    "paths": {
...
        "/passenger/login": {
            "post": {
                "description": "Initiates login by sending a verification code to the passenger's phone number. Checks if the phone is registered and logs the code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passenger"
                ],
                "summary": "Initiate passenger login",
                "parameters": [
                    {
                        "description": "Passenger login details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/passenger/register": {
            "post": {
                "description": "Registers a new passenger with a full name and phone number, sending a verification code. Validates input, checks for duplicate phone numbers, and logs the code (replace with SMS in production).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passenger"
                ],
                "summary": "Register a new passenger",
                "parameters": [
                    {
                        "description": "Passenger registration details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PassengerRegisterRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/protected": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns user information for an authenticated user. Requires a valid JWT token in the Authorization header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Access protected resource",
                "responses": {}
            }
        },
...
        "/protected/taxist-rating/{taxist_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a taxist's rating by taxist_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Update taxist rating",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Taxist ID",
                        "name": "taxist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Taxist rating details",
                        "name": "taxist_rating",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TaxistRating"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/taxist/login": {
            "post": {
                "description": "Initiates login by sending a verification code to the taxist's phone number. Checks if the phone is registered and logs the code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Taxist"
                ],
                "summary": "Initiate taxist login",
                "parameters": [
                    {
                        "description": "Taxist login details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/taxist/register": {
            "post": {
                "description": "Registers a new taxist with personal and vehicle details, sending a verification code. Validates input, checks for duplicate phone numbers, and logs the code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Taxist"
                ],
                "summary": "Register a new taxist",
                "parameters": [
                    {
                        "description": "Taxist registration details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TaxistRegisterRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/verify": {
            "post": {
                "description": "Verifies a phone number using a 4-digit code for passengers or taxists, issuing a JWT token upon success. Deletes the used code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passenger",
                    "Taxist"
                ],
                "summary": "Verify phone number",
                "parameters": [
                    {
                        "description": "Verification details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.VerifyCodeRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "+12345678901"
                }
            }
        },
...
        "handlers.PassengerRegisterRequest": {
            "type": "object",
            "required": [
                "full_name",
                "phone"
            ],
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+12345678901"
                }
            }
        },
...
        "handlers.TaxistRating": {
            "type": "object",
            "required": [
                "rating"
            ],
            "properties": {
                "rating": {
                    "type": "number",
                    "example": 4
                }
            }
        },
        "handlers.TaxistRegisterRequest": {
            "type": "object",
            "required": [
                "car_make",
                "car_model",
                "car_number",
                "car_year",
                "full_name",
                "phone"
            ],
            "properties": {
                "car_make": {
                    "type": "string",
                    "example": "Toyota"
                },
                "car_model": {
                    "type": "string",
                    "example": "Camry"
                },
                "car_number": {
                    "type": "string",
                    "example": "ABC123"
                },
                "car_year": {
                    "type": "integer",
                    "maximum": 2026,
                    "minimum": 1900,
                    "example": 2020
                },
                "full_name": {
                    "type": "string",
                    "example": "Jane Smith"
                },
                "phone": {
                    "type": "string",
                    "example": "+12345678902"
                }
            }
        },
        "handlers.VerifyCodeRequest": {
            "type": "object",
            "required": [
                "code",
                "phone"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "1234"
                },
                "phone": {
                    "type": "string",
                    "example": "+12345678901"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}